keygen/requirements.txt
click==8.1.3
pyjwt==2.6.0
